send chat inference config when sending message
