+++ src/data/chats/chat.rs
@@ -54,6 +54,27 @@ struct ChatData {
 }
 
 #[derive(Debug)]
+pub struct ChatInferenceParams {
+    pub frequency_penalty: f32,
+    pub max_tokens: u32,
+    pub presence_penalty: f32,
+    pub temperature: f32,
+    pub top_p: f32,
+}
+
+impl Default for ChatInferenceParams {
+    fn default() -> Self {
+        Self {
+            frequency_penalty: 0.0,
+            max_tokens: 2048,
+            presence_penalty: 0.0,
+            temperature: 1.0,
+            top_p: 1.0,
+        }
+    }
+}
+
+#[derive(Debug)]
 pub struct Chat {
     /// Unix timestamp in ms.
     pub id: ChatID,
@@ -175,7 +196,13 @@ impl Chat {
             }
         }
     }
-    pub fn send_message_to_model(&mut self, prompt: String, loaded_file: &File, backend: &Backend) {
+    pub fn send_message_to_model(
+        &mut self,
+        prompt: String,
+        inference_params: &ChatInferenceParams,
+        loaded_file: &File,
+        backend: &Backend,
+    ) {
         let (tx, rx) = channel();
         let mut messages: Vec<_> = self
             .messages
@@ -193,20 +220,21 @@ impl Chat {
             name: None,
         });
 
+        let ip = inference_params;
         let cmd = Command::Chat(
             ChatRequestData {
                 messages,
                 model: self.model_filename.clone(),
-                frequency_penalty: None,
+                frequency_penalty: Some(ip.frequency_penalty),
                 logprobs: None,
                 top_logprobs: None,
-                max_tokens: None,
-                presence_penalty: None,
+                max_tokens: Some(ip.max_tokens),
+                presence_penalty: Some(ip.presence_penalty),
                 seed: None,
                 stop: None,
                 stream: Some(true),
-                temperature: None,
-                top_p: None,
+                temperature: Some(ip.temperature),
+                top_p: Some(ip.top_p),
                 n: None,
                 logit_bias: None,
             },
